Move setting error icon to the left of section rows
This change moves the error icon for a NodeEditor section to the left of the title. Previously this icon would be to the right of the visibility indicator. This would create a poor UX experience when the user would toggle a topic on which could result in an error on the topic immediately displacing the visibility icon they clicked and placing the error icon under their cursor. After this change the visibility icon remains under the cursor even if there is a topic error.