Remember the dataset range for IndexDatasetsBuilder
When the builder receives state with no update it should still return the dataset range so that the rendering update continues to use the range for the dataset rather than falling back to automatic bounds for chartjs. This avoid having "gaps" at the start or end of the plot and sizing the plot to the data.