Reduce instances of erroneous "reset view" on the plot
The plot and state transition panels use different approaches to "global bounds" when using the "sync with other plots" feature. The plot panel only sets the global bounds when a user interacts with the panel (zoom or pan), otherwise it leaves the bounds unset. The state transition panel sets the bounds when it loads. When both panels are used on a layout, this causes the existing plot panel logic for when to show the reset button to think the bounds are set and you could reset them. However, the bounds already match what the panel is displaying so there's no need to reset.

This change updates the "show reset" logic to account for bounds being set and the panel already matching those bounds. When that happens, there's no need to show the reset button.