Move setSeries to an update action
The TimestampDatasetsBuilder and CustomDatasetsBuilder need to send the series configuration to their respective workers. Prior to this change this was done with a call to `setSeries` on the exposed RPC interface with a _void_ since this call happened in an async function that did not wait. This change moves the call into an update action so it can be dispatched with an await like the other pending update actions and be captured by whatever error mechanism captures the async function containing the dispatched calls.